Begin implementing statistics panel
