+++ src/components/editor/Editor.vue
@@ -12,26 +12,25 @@
     <section class="body">
       <Paper :graph="graph" v-on:send-element="receiveElement"></Paper>
       <section class="additional-info">
-        <section class="statistics">
-          <!-- Statistics will go here -->
-          <span>Statistics</span>
-        </section>
-
-        <Preview></Preview>
+        <Statistics :sql="sql"></Statistics>
+        <Preview :sql="sql"></Preview>
       </section>
     </section>
   </section>
 </template>
 
 <script>
+import { createSQL } from '../../util/sql_util.js'
 import { createGraph, getElementName } from '../../util/jointjs_util'
 import { RECEIVE_GRAPH, UPDATE_GRAPH } from '../../store/mutation_types'
 import Paper from './Paper'
 import Preview from './Preview'
+import Statistics from './Statistics'
 export default {
   components: {
     Paper,
-    Preview
+    Preview,
+    Statistics
   },
   data: function () {
     return {
@@ -40,6 +39,10 @@ export default {
     }
   },
   computed: {
+    sql: function () {
+      let json = this.$store.state.graphJSON
+      return createSQL(json)
+    },
     currentElementName: {
       get: function () {
         return getElementName(this.currentElement)
@@ -100,10 +103,5 @@ export default {
 
   .additional-info {
     display: flex;
-  }
-
-  .additional-info > section {
-    border: 1px solid black;
-    flex: 1;
   }
 </style>
