+++ src/components/editor/Statistics.vue
@@ -1,0 +1,18 @@
+<template lang="html">
+  <section class="statistics">
+    <!-- Statistics will go here -->
+    <span>Statistics</span>
+  </section>
+</template>
+
+<script>
+export default {
+}
+</script>
+
+<style lang="css">
+  .statistics {
+    border: 1px solid black;
+    flex: 1;
+  }
+</style>
